split checksum into a separated file and fix comment
