+++ lightning/restore/checksum_test.go
@@ -1,0 +1,200 @@
+package restore
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+	"sync"
+	"time"
+
+	pd "github.com/tikv/pd/client"
+
+	"github.com/DATA-DOG/go-sqlmock"
+	. "github.com/pingcap/check"
+	"github.com/pingcap/errors"
+
+	. "github.com/pingcap/tidb-lightning/lightning/checkpoints"
+)
+
+var _ = Suite(&checksumSuite{})
+
+type checksumSuite struct{}
+
+func MockDoChecksumCtx(db *sql.DB) context.Context {
+	ctx := context.Background()
+	manager := newTiDBChecksumExecutor(db)
+	return context.WithValue(ctx, &checksumManagerKey, manager)
+}
+
+func (s *checksumSuite) TestDoChecksum(c *C) {
+	db, mock, err := sqlmock.New()
+	c.Assert(err, IsNil)
+
+	mock.ExpectQuery("\\QSELECT VARIABLE_VALUE FROM mysql.tidb WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WillReturnRows(sqlmock.NewRows([]string{"VARIABLE_VALUE"}).AddRow("10m"))
+	mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WithArgs("100h0m0s").
+		WillReturnResult(sqlmock.NewResult(1, 1))
+	mock.ExpectQuery("\\QADMIN CHECKSUM TABLE `test`.`t`\\E").
+		WillReturnRows(
+			sqlmock.NewRows([]string{"Db_name", "Table_name", "Checksum_crc64_xor", "Total_kvs", "Total_bytes"}).
+				AddRow("test", "t", 8520875019404689597, 7296873, 357601387),
+		)
+	mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WithArgs("10m").
+		WillReturnResult(sqlmock.NewResult(2, 1))
+	mock.ExpectClose()
+
+	ctx := MockDoChecksumCtx(db)
+	checksum, err := DoChecksum(ctx, db, &TidbTableInfo{DB: "test", Name: "t"})
+	c.Assert(err, IsNil)
+	c.Assert(*checksum, DeepEquals, RemoteChecksum{
+		Schema:     "test",
+		Table:      "t",
+		Checksum:   8520875019404689597,
+		TotalKVs:   7296873,
+		TotalBytes: 357601387,
+	})
+
+	c.Assert(db.Close(), IsNil)
+	c.Assert(mock.ExpectationsWereMet(), IsNil)
+}
+
+func (s *checksumSuite) TestDoChecksumParallel(c *C) {
+	db, mock, err := sqlmock.New()
+	c.Assert(err, IsNil)
+
+	mock.ExpectQuery("\\QSELECT VARIABLE_VALUE FROM mysql.tidb WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WillReturnRows(sqlmock.NewRows([]string{"VARIABLE_VALUE"}).AddRow("10m"))
+	mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WithArgs("100h0m0s").
+		WillReturnResult(sqlmock.NewResult(1, 1))
+	for i := 0; i < 5; i++ {
+		mock.ExpectQuery("\\QADMIN CHECKSUM TABLE `test`.`t`\\E").
+			WillDelayFor(100 * time.Millisecond).
+			WillReturnRows(
+				sqlmock.NewRows([]string{"Db_name", "Table_name", "Checksum_crc64_xor", "Total_kvs", "Total_bytes"}).
+					AddRow("test", "t", 8520875019404689597, 7296873, 357601387),
+			)
+	}
+	mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WithArgs("10m").
+		WillReturnResult(sqlmock.NewResult(2, 1))
+	mock.ExpectClose()
+
+	ctx := MockDoChecksumCtx(db)
+
+	// db.Close() will close all connections from its idle pool, set it 1 to expect one close
+	db.SetMaxIdleConns(1)
+	var wg sync.WaitGroup
+	wg.Add(5)
+	for i := 0; i < 5; i++ {
+		go func() {
+			defer wg.Done()
+			checksum, err := DoChecksum(ctx, db, &TidbTableInfo{DB: "test", Name: "t"})
+			c.Assert(err, IsNil)
+			c.Assert(*checksum, DeepEquals, RemoteChecksum{
+				Schema:     "test",
+				Table:      "t",
+				Checksum:   8520875019404689597,
+				TotalKVs:   7296873,
+				TotalBytes: 357601387,
+			})
+		}()
+	}
+	wg.Wait()
+
+	c.Assert(db.Close(), IsNil)
+	c.Assert(mock.ExpectationsWereMet(), IsNil)
+}
+
+func (s *checksumSuite) TestIncreaseGCLifeTimeFail(c *C) {
+	db, mock, err := sqlmock.New()
+	c.Assert(err, IsNil)
+
+	for i := 0; i < 5; i++ {
+		mock.ExpectQuery("\\QSELECT VARIABLE_VALUE FROM mysql.tidb WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+			WillReturnRows(sqlmock.NewRows([]string{"VARIABLE_VALUE"}).AddRow("10m"))
+		mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+			WithArgs("100h0m0s").
+			WillReturnError(errors.Annotate(context.Canceled, "update gc error"))
+	}
+	// This recover GC Life Time SQL should not be executed in DoChecksum
+	mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WithArgs("10m").
+		WillReturnResult(sqlmock.NewResult(1, 1))
+	mock.ExpectClose()
+
+	ctx := MockDoChecksumCtx(db)
+	var wg sync.WaitGroup
+	wg.Add(5)
+	for i := 0; i < 5; i++ {
+		go func() {
+			_, err = DoChecksum(ctx, db, &TidbTableInfo{DB: "test", Name: "t"})
+			c.Assert(err, ErrorMatches, "update GC lifetime failed: update gc error: context canceled")
+			wg.Done()
+		}()
+	}
+	wg.Wait()
+
+	_, err = db.Exec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E", "10m")
+	c.Assert(err, IsNil)
+
+	c.Assert(db.Close(), IsNil)
+	c.Assert(mock.ExpectationsWereMet(), IsNil)
+}
+
+func (s *checksumSuite) TestDoChecksumWithErrorAndLongOriginalLifetime(c *C) {
+	db, mock, err := sqlmock.New()
+	c.Assert(err, IsNil)
+
+	mock.ExpectQuery("\\QSELECT VARIABLE_VALUE FROM mysql.tidb WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WillReturnRows(sqlmock.NewRows([]string{"VARIABLE_VALUE"}).AddRow("300h"))
+	mock.ExpectQuery("\\QADMIN CHECKSUM TABLE `test`.`t`\\E").
+		WillReturnError(errors.Annotate(context.Canceled, "mock syntax error"))
+	mock.ExpectExec("\\QUPDATE mysql.tidb SET VARIABLE_VALUE = ? WHERE VARIABLE_NAME = 'tikv_gc_life_time'\\E").
+		WithArgs("300h").
+		WillReturnResult(sqlmock.NewResult(1, 1))
+	mock.ExpectClose()
+
+	ctx := MockDoChecksumCtx(db)
+	_, err = DoChecksum(ctx, db, &TidbTableInfo{DB: "test", Name: "t"})
+	c.Assert(err, ErrorMatches, "compute remote checksum failed: mock syntax error.*")
+
+	c.Assert(db.Close(), IsNil)
+	c.Assert(mock.ExpectationsWereMet(), IsNil)
+}
+
+type testPDClient struct {
+	pd.Client
+}
+
+func (c *testPDClient) UpdateServiceGCSafePoint(ctx context.Context, serviceID string, ttl int64, safePoint uint64) (uint64, error) {
+	return 0, nil
+}
+
+func (s *checksumSuite) TestGcTTLManager(c *C) {
+	manager := gcTTLManager{pdClient: &testPDClient{}}
+	ctx := context.Background()
+
+	for i := uint64(1); i <= 5; i++ {
+		err := manager.addOneJob(ctx, fmt.Sprintf("test%d", i), i)
+		c.Assert(err, IsNil)
+		c.Assert(manager.currentTs, Equals, uint64(1))
+	}
+
+	manager.removeOneJob("test2")
+	c.Assert(manager.currentTs, Equals, uint64(1))
+
+	manager.removeOneJob("test1")
+	c.Assert(manager.currentTs, Equals, uint64(3))
+
+	manager.removeOneJob("test3")
+	c.Assert(manager.currentTs, Equals, uint64(4))
+
+	manager.removeOneJob("test4")
+	c.Assert(manager.currentTs, Equals, uint64(5))
+
+	manager.removeOneJob("test5")
+	c.Assert(manager.currentTs, Equals, uint64(0))
+}
